requirements.txt
-i https://pypi.org/simple
-e git+https://github.com/twintproject/twint.git@ad27650fbc0bf8c3f2c78449088a5ede7239f53a#egg=twint
aiodns==2.0.0
aiohttp-socks==0.2.2
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
cchardet==2.1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.0
cycler==0.10.0
decorator==4.4.0
dnspython==1.16.0
elasticsearch==7.0.2
eventlet==0.25.0
fake-useragent==0.1.11
flask-cors==3.0.8
flask==1.1.0
geographiclib==1.49
geopy==1.20.0
greenlet==0.4.15
gunicorn==19.9.0
idna-ssl==1.1.0 ; python_version < '3.7'
idna==2.8
imageio==2.5.0
itsdangerous==1.1.0
jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
markupsafe==1.1.1
monotonic==1.5
multidict==4.5.2
networkx==2.3
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
pillow==6.1.0
pycares==3.0.0
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.0
pysocks==1.7.0
pytesseract==0.2.7
python-dateutil==2.8.0
pytz==2019.1
pywavelets==1.0.3
regex==2019.6.8
requests==2.22.0
schedule==0.6.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
soupsieve==1.9.2
typing-extensions==3.7.4 ; python_version < '3.7'
typing==3.7.4 ; python_version < '3.7'
urllib3==1.25.3
werkzeug==0.15.4
yapf==0.27.0
yarl==1.3.0
